pyproject.toml
[project]
name = "agg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agent>=0.1.0",
]
